indoor_nav/requirements.txt
Django==2.2.9
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
